example: rename files to initiator, responder and common
