+++ examples/lakers-nrf52840/src/bin/common.rs
